+++ crates/config/src/aggregator_config.rs
@@ -1,0 +1,6 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub struct AggregatorConfig {
+    pub ip_address: String,
+}
