+++ crates/config/src/el_config.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub struct ELConfig {
+    pub registry_coordinator_addr: String,
+    pub operator_state_retriever_addr: String,
+    pub delegation_manager_addr: String,
+    pub avs_directory_addr: String,
+    pub strategy_manager_addr: String,
+}
