+++ crates/config/src/metrics_config.rs
@@ -1,0 +1,6 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub struct MetricsConfig {
+    pub port_address: String,
+}
