+++ crates/config/src/node_config.rs
@@ -1,0 +1,6 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub struct NodeConfig {
+    pub node_port_address: String,
+}
