+++ crates/config/src/task_manager.rs
@@ -1,0 +1,6 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub struct TaskManagerConfig {
+    pub signer: String,
+}
